requirements.txt
aioquic
anyio
attrs
botocore
brotli
certifi
cffi
chardet
click
colorama
cryptography
curio
cycler
cython
dask
python-dateutil
defusedxml
dnspython
docopt
docutils
fastparquet
filelock
fsspec
google
ipython
lxml
matplotlib
numpy
pyopenssl
pandas
pillow
pyinstaller
pyyaml
requests
scipy
scikit-learn
sqlalchemy
trio
whois
openpyxl

shodan
selenium
beautifulsoup4
tqdm
pandas
selenium
requests
